Moved ujit code invalidation test into bootstraptest
